create openMultiCameras.cpp
allows user to open two cameras simultaneously. The code was executed on Ubuntu 15.10. The user should also be aware that each camera was connected to their own USB Host Controller to avoid bandwidth data overload. 

Updating the code to three or four or arbitrary number of cameras is simply creating more videocapture objects, mat frame objects, creating multiple windows for the added cameras and showing the frames in the appropriate camera window.