create videoStiching.cpp
First draft of code uses OpenCV Thread Building Block API a.k.a multithreading programming to interact with multiple video files. For each video file, there is a designated thread that starts video file capturing process and captures video frames. Currently we have 4 video files that are inputted to the program, so their is a thread to manage each video file. In total there are 4 threads. The goal is for the program to support an arbitrary number of video files.

Once 4 video files that are 2 minutes each have been stitched into a single 360 video panorama video file, next is to test against real-time video cameras.