fixed videoStitching.cpp
program now displays the correct feature keypoints per frame for each video file or camera
I removed the vector<ImageFeatures> features and replaced with a vector<concurrent_queue<ImageFeatures>*> featuresQueue. This concurrent_queue<ImageFeatures>* is a pointer called featuresQueue which points to a queue that holds the features for each frame within a video file. The vector<concurrent_queue<ImageFeatures*> is a vector of pointer of featuresQueue objects that point to each camera/video and their frames features